+++ src/main/java/carrenting/server/ICarRenting.java
@@ -18,7 +18,7 @@ public interface ICarRenting extends Remote  {
 	public ArrayList<Car> getCars(String garage) throws RemoteException;
 	public Car getCar(String numPlate) throws RemoteException;
 	public ArrayList<Rent> getRents() throws RemoteException;
-	public void storeCar(String garage, String numberPlate, String brand, String model, int pricePerDay) throws RemoteException;
+	public void storeCar(String garage, String numberPlate, String brand, String model, double pricePerDay) throws RemoteException;
 	public void deleteCar(String numberPlate) throws RemoteException;
 	public void deleteGarage(String garage) throws RemoteException;
 }
